✨ tweak: move the askai logic to message event file
