+++ src/bot.js
@@ -6,9 +6,6 @@ const {
 	Partials,
 } = require("discord.js");
 const {
-	sendEmbedMessage,
-	CloseButtonComponent,
-	FeedbackButtonComponent,
 	serverTime,
 } = require("./utils/core");
 
@@ -22,8 +19,6 @@ require("dotenv").config();
 const {
 	DISCORD_BOT_TOKEN,
 	DISCORD_SUPPORT_ROLE_ID,
-	CONTEXT_ID,
-	ASKAI_CHANNEL,
 	REDIS_SERVER_URL
 } = process.env;
 
@@ -44,78 +39,8 @@ const redis = new Redis(REDIS_SERVER_URL);
 const context = new ContextSDK({});
 
 // listen to post messages
-client.on("messageCreate", async (message) => {
+// client.on("messageCreate", async (message) => {
 
-	// respond to ask command
-	if ((message.content.startsWith('!askai') || message.content.startsWith('!ask'))) {
-		let question = message.content.startsWith('!askai') ? message.content.slice(6) : message.content.slice(4);
-		const gettingStartedASKAI = `Hello, kindly use \`!ask\` or \`!askai\` followed by your question to get started.`;
-
-		// check if there's a question, if not, send the getting started message, if there's a question, send the response
-		if (!question) {
-			message.reply({
-				content: `Hey <@${message.author.id}> 👇`,
-				embeds: [sendEmbedMessage(gettingStartedASKAI)],
-			});
-		} else {
-			if (message.channel.id === ASKAI_CHANNEL) {
-				let aiMessageLoading = await message.channel.send({
-					embeds: [
-						sendEmbedMessage("**🤖 Beep Boop Boop Beep:** " + `<a:load:1210497921158619136> thinking...`),
-					],
-				});
-
-				await context.query({
-					botId: CONTEXT_ID,
-					query: question,
-					onComplete: async (query) => {
-						// respond to the user with the answer from the AI
-						await message.channel.messages.fetch(aiMessageLoading.id).then((msg) => {
-							msg.edit({
-								content: `Hey <@${message.author.id}> 👇`,
-								embeds: [
-									sendEmbedMessage(`**Response:**\n${query.output.toString()}`),
-								],
-								components: [FeedbackButtonComponent()],
-
-							})
-							redis.set(msg.id, query._id);
-						}
-						);
-
-					},
-					onError: async (error) => {
-						console.error(error);
-
-						// send a message indicates unseccesful response from the AI
-						await message.channel.messages.fetch(aiMessageLoading.id).then((msg) =>
-							msg.edit({
-								content: `Hey <@${message.author.id}> 👇`,
-								embeds: [
-									sendEmbedMessage(`**Response:**\nI'm sorry, I couldn't find a response to your question. Please try again later.`),
-								],
-
-							})
-						);
-
-					},
-				});
-
-			} else {
-				// if the command is not from the channel
-				message.reply({
-					content: `Hey <@${message.author.id}> 👇`,
-					embeds: [sendEmbedMessage(`You can ask me all things thirdweb in the <#${ASKAI_CHANNEL}> channel. Just type your question after the command \`!askai\` or \`!ask\` to get started.`)],
-				});
-			}
-		}
-	}
-
-});
-
-//listen to button clicks
-// client.on("interactionCreate", async (interaction) => {
-	
 // });
 
 /**
